Migrate to new ALC structure and support WinPS (#6)
Migrates to a new Assembly Load Context structure which ensures the
dependencies used are the ones shipped with the module. It also ensures
that no dependency might accidentially escape the ALC by loading it
explicitly inside the ALC.

This also adds limited support for Windows PowerShell 5.1. While WinPS
does not support ALCs it can load assemblies side by side removing some
of the need for an ALC. There are no guarantees that this will always
load our version only that dotnet can load an assembly that matches our
required name if present.