Handle more integer types
Add support for more integer types like BigInteger and Int128/UInt128.
Also slightly changes the serialization behaviour of a UInt64 that is
larger than Int64.MaxValue to be serialized as a string to avoid casting
the value to the signed variant.

Changes the string casting behaviour to use PowerShell's ETS which
supports more complex string conversions than just ToString().